requirements.txt
faiss-cpu
numpy
pandas
sentence-transformers
openai
python-dotenv